refactor: move init_app to module initialization
temporarily move Database class to separate module to avoid circular
import
